requirements.txt
# Finans Asistanı v11 - Gereksinimler

# API & AI
groq>=0.4.0
python-dotenv>=1.0.0

# Web Scraping (Fallback)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Finans Verileri
yfinance>=0.2.36
tefas-crawler>=0.3.0

# Browser Automation
playwright>=1.40.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Uyarı bastırma
urllib3>=2.0.0
